Introduce `Design2CommonRoundedBorderRadiusStyle` and apply to `ContentPageBanner` component
