+++ src/components/shared/Design2_CommonUtils.js
@@ -1,0 +1,20 @@
+/**
+ * Design2_CommonUtils.js
+ * Chewbaaka
+ *
+ * Author   : Tomiko
+ * Created  : Jun 09, 2022
+ * Updated  : Jun 09, 2022
+ */
+
+import { join } from './Utils'
+
+const __Design2CommonRoundedBorderRadiusStyle = [
+  "CommonRoundedBorderRadiusStyle"
+]
+
+const Design2CommonRoundedBorderRadiusStyle = join(__Design2CommonRoundedBorderRadiusStyle)
+
+export {
+  Design2CommonRoundedBorderRadiusStyle
+}
