+++ src/components/shared/Design2_ImageViewFamilyUtils.js
@@ -7,6 +7,8 @@
  * Updated  : Jun 09, 2022
  */
 
+import { join } from './Utils'
+
 const __Design2ImageViewFamilyImgStyles = [
   "ImageViewFamilyImgLayoutStyles",
   "ImageViewFamilyImgBorderStyles"
@@ -22,23 +24,14 @@ const __Design2ImageViewFamilyCaptionContainerStyles = [
   "ImageViewFamilyCaptionContainerStyles"
 ]
 
-function join(arr) {
-  return arr.join(" ")
-}
-
 const Design2ImageViewFamilyImgStyles = join(__Design2ImageViewFamilyImgStyles)
 
 const Design2ImageViewFamilyContainerStyles = join(__Design2ImageViewFamilyContainerStyles)
 
 const Design2ImageViewFamilyCaptionContainerStyles = join(__Design2ImageViewFamilyCaptionContainerStyles)
 
-function combineStyles(a, b) {
-  return (a + " " + b);
-}
-
 export {
   Design2ImageViewFamilyImgStyles,
   Design2ImageViewFamilyContainerStyles,
-  Design2ImageViewFamilyCaptionContainerStyles,
-  combineStyles
+  Design2ImageViewFamilyCaptionContainerStyles
 }
