+++ src/components/shared/Utils.js
@@ -1,0 +1,21 @@
+/**
+ * Utils.js
+ * Chewbaaka
+ *
+ * Author   : Tomiko
+ * Created  : Jun 09, 2022
+ * Updated  : Jun 09, 2022
+ */
+
+function join(arr) {
+  return arr.join(" ")
+}
+
+function combineStyles(a, b) {
+  return (a + " " + b);
+}
+
+export {
+  join,
+  combineStyles
+}
